requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
Django==3.0.8
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==14.1.0
django-bootstrap4==2.3.1
django-crispy-forms==1.9.2
django-filter==2.4.0
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.2.0
psycopg2-binary==2.8.6
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
whitenoise==5.2.0
